chore: Split governance type in separate file for easier navigation
